feat: creating restTemplate requesto to CambioMicroservice, refactored convertedValue type

+++ book-service/src/main/java/com/andoliver46/ms/models/BookModel.java
@@ -1,6 +1,7 @@
 package com.andoliver46.ms.models;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Objects;
 
@@ -28,7 +29,7 @@ public class BookModel implements Serializable{
 	@Temporal(TemporalType.DATE)
 	private Date launchDate;
 	@Column(nullable = false)
-	private Double price;
+	private BigDecimal price;
 	@Column(nullable = false, length = 250)
 	private String title;
 	@Transient
@@ -38,7 +39,7 @@ public class BookModel implements Serializable{
 	
 	public BookModel() {}
 	
-	public BookModel(Long id, String author, String title, Date launchDate, Double price, String currency,
+	public BookModel(Long id, String author, String title, Date launchDate, BigDecimal price, String currency,
 			String environment) {
 		super();
 		this.id = id;
@@ -74,11 +75,11 @@ public class BookModel implements Serializable{
 		this.launchDate = launchDate;
 	}
 
-	public Double getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
 
-	public void setPrice(Double price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
